further refactor allowing ops to be used with current api
