+++ client/client_tests/util/test_milvus_util.py
@@ -4,8 +4,6 @@
 
 import pytest
 from nv_ingest_client.util.milvus import (
-    MilvusOperator,
-    _dict_to_params,
     create_nvingest_collection,
     grab_meta_collection_info,
     reconstruct_pages,
@@ -13,39 +11,42 @@ from nv_ingest_client.util.milvus import (
     pandas_file_reader,
     create_nvingest_index_params,
 )
+from nv_ingest_client.util.vdb.milvus import Milvus, _dict_to_params
 from nv_ingest_client.util.util import ClientConfigSchema
 import pandas as pd
+import unittest
+from nv_ingest_client.util.vdb.adt_vdb import VDB
 
 
 @pytest.fixture
 def milvus_test_dict():
-    mil_op = MilvusOperator()
-    kwargs = mil_op.milvus_kwargs
+    mil_op = Milvus()
+    kwargs = mil_op.__dict__
     kwargs["collection_name"] = mil_op.collection_name
     return kwargs
 
 
 def test_extra_kwargs(milvus_test_dict):
-    mil_op = MilvusOperator(filter_errors=True)
-    assert "filter_errors" in mil_op.milvus_kwargs
-    assert mil_op.milvus_kwargs["filter_errors"] is True
+    mil_op = Milvus(filter_errors=True)
+    assert "filter_errors" in mil_op.__dict__["kwargs"]
+    assert mil_op.__dict__["kwargs"]["filter_errors"] is True
 
 
 @pytest.mark.parametrize("collection_name", [None, "name"])
 def test_op_collection_name(collection_name):
     if collection_name:
-        mo = MilvusOperator(collection_name=collection_name)
+        mo = Milvus(collection_name=collection_name)
     else:
         # default
         collection_name = "nv_ingest_collection"
-        mo = MilvusOperator()
+        mo = Milvus()
     cr_collection_name, conn_params = mo.get_connection_params()
     wr_collection_name, write_params = mo.get_write_params()
     assert cr_collection_name == wr_collection_name == collection_name
 
 
 def test_op_connection_params(milvus_test_dict):
-    mo = MilvusOperator()
+    mo = Milvus()
     cr_collection_name, conn_params = mo.get_connection_params()
     assert cr_collection_name == milvus_test_dict["collection_name"]
     for k, v in conn_params.items():
@@ -53,7 +54,7 @@ def test_op_connection_params(milvus_test_dict):
 
 
 def test_op_write_params(milvus_test_dict):
-    mo = MilvusOperator()
+    mo = Milvus()
     collection_name, wr_params = mo.get_write_params()
     assert collection_name == milvus_test_dict["collection_name"]
     for k, v in wr_params.items():
@@ -70,7 +71,7 @@ def test_op_write_params(milvus_test_dict):
     ],
 )
 def test_op_dict_to_params(collection_name, expected_results):
-    mo = MilvusOperator()
+    mo = Milvus()
     _, wr_params = mo.get_write_params()
     response = _dict_to_params(collection_name, wr_params)
     if isinstance(collection_name, str):
@@ -381,3 +382,61 @@ def test_create_nvingest_index_params(sparse, gpu_index, gpu_search, local_index
     assert (
         actual_params == expected_params
     ), f"Index parameters do not match expected values.\nActual: {actual_params}\nExpected: {expected_params}"
+
+
+class MockVDB(VDB):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.run_called = False
+        self.run_records = None
+
+    def create_index(self, **kwargs):
+        pass
+
+    def write_to_index(self, records: list, **kwargs):
+        pass
+
+    def retrieval(self, queries: list, **kwargs):
+        pass
+
+    def reindex(self, records: list, **kwargs):
+        pass
+
+    def run(self, records):
+        self.run_called = True
+        self.run_records = records
+
+
+class TestVDBUpload(unittest.TestCase):
+    def setUp(self):
+        self.mock_vdb = MockVDB()
+        self.test_records = [
+            {
+                "metadata": {
+                    "content": "test content",
+                    "embedding": [0.1, 0.2, 0.3],
+                    "source_metadata": {"source": "test.pdf"},
+                    "content_metadata": {"type": "text"},
+                }
+            }
+        ]
+
+    def test_vdb_upload_uses_adt(self):
+        """Test that when a VDB operator is supplied, its run method is called with the records."""
+        from nv_ingest_client.client.interface import Ingestor
+
+        # Create an Ingestor instance
+        ingestor = Ingestor()
+        # Add the mock VDB operator
+        ingestor.vdb_upload(vdb_op=self.mock_vdb)
+        assert ingestor._vdb_bulk_upload == self.mock_vdb
+
+    def test_vdb_upload_invalid_op(self):
+        """Test that an invalid VDB operator raises a ValueError."""
+        from nv_ingest_client.client.interface import Ingestor
+
+        ingestor = Ingestor()
+
+        # Try to add an invalid VDB operator
+        with self.assertRaises(ValueError):
+            ingestor.vdb_upload(vdb_op="not_a_vdb_op")
