+++ client/src/nv_ingest_client/util/vdb/__init__.py
@@ -1,0 +1,7 @@
+from nv_ingest_client.util.vdb.adt_vdb import VDB
+from nv_ingest_client.util.vdb.milvus import Milvus
+
+available_vdb_ops = {
+    "milvus": Milvus,
+}
+__all__ = ["VDB", "available_vdb_ops"]
